LFS:  stage-3-build_basic_system/gcc.sh updated

+++ stage-1-lfs-stable-x86_64/stage-3-build_basic_system/gcc.sh
@@ -11,10 +11,10 @@ source "${ROOT}unpack_source_archive.sh" "${PRGNAME}" || exit 1
 
 TMP_DIR="/tmp/pkg-${PRGNAME}-${VERSION}"
 rm -rf "${TMP_DIR}"
-mkdir -pv "${TMP_DIR}"/{lib,usr/lib/bfd-plugins,usr/share/gdb/auto-load/usr/lib}
+mkdir -pv "${TMP_DIR}"/{usr/lib/bfd-plugins,usr/share/gdb/auto-load/usr/lib}
 
-# установим имя каталога для 64-битных библиотек по умолчанию как 'lib'
-sed -e '/m64=/s/lib64/lib/' -i.orig gcc/config/i386/t-linux64
+# изменим каталог для установки библиотек с lib64 на lib
+sed -e '/m64=/s/lib64/lib/' -i.orig "${PRGNAME}/config/i386/t-linux64" || exit 1
 
 # документация gcc рекомендует собирать gcc в отдельном каталоге
 mkdir build
@@ -25,18 +25,22 @@ cd build || exit 1
 # сообщим GCC, что нужно ссылаться на установленную в системе библиотеку Zlib,
 # а не на собственную внутреннюю копию
 #    --with-system-zlib
-../configure            \
-    --prefix=/usr       \
-    LD=ld               \
-    --disable-multilib  \
-    --disable-bootstrap \
-    --with-system-zlib  \
+../configure             \
+    --prefix=/usr        \
+    LD=ld                \
+    --enable-default-pie \
+    --enable-default-ssp \
+    --disable-multilib   \
+    --disable-bootstrap  \
+    --with-system-zlib   \
     --enable-languages=c,c++ || exit 1
 
 make || make -j1 || exit 1
 
-#### Набор тестов для GCC на данном этапе считается критическим. Нельзя
-# пропускать его ни при каких обстоятельствах
+###
+# Набор тестов для GCC на данном этапе считается критическим. Нельзя пропускать
+# его ни при каких обстоятельствах
+###
 #
 # известно, что один набор тестов в наборе тестов GCC переполняет стек, поэтому
 # увеличим размер стека
@@ -53,11 +57,8 @@ make || make -j1 || exit 1
 # echo "${JUNK}" > /dev/null
 # su tester -c "PATH=$PATH make -k check"
 
-# пишем результаты тестов GCC в gcc-test.log. Известно, что шесть тестов,
-# связанных с get_time не проходят (по-видимому, они связаны с локалью en_HK).
-# Тесты asan_test.C, co-ret-17-void-ret-coro.C, pr95519-05-gro.C, pr80166.c так
-# же не проходят с glibc-2.32
-# ../contrib/test_summary 2>&1 | grep -A7 Summ > gcc-test.log
+# пишем результаты тестов GCC в gcc-test.log
+# ../contrib/test_summary 2>&1 | grep -A7 Summ > "${PRGNAME}-test.log"
 
 # вернем владельца сборочной директории обратно
 # chown -Rv root:root .
@@ -65,52 +66,46 @@ make || make -j1 || exit 1
 # установим пакет
 make install DESTDIR="${TMP_DIR}"
 
-# удалим ненужную директорию
-# /usr/lib/gcc/x86_64-pc-linux-gnu/${VERSION}/include-fixed/bits
-rm -rf \
-    "${TMP_DIR}/usr/lib/gcc/$(gcc -dumpmachine)/${VERSION}/include-fixed/bits/"
-
-# создадим символическую ссылку в /lib/
-# cpp -> ../usr/bin/cpp
-# требуемую FHS по историческим причинам
-(
-    cd "${TMP_DIR}/lib" || exit 1
-    ln -sv ../usr/bin/cpp cpp
-)
-
+# создадим символическую ссылку в /usr/lib, требуемую FHS по "историческим"
+# причинам
+#    cpp -> /usr/bin/cpp
 # многие программы используют имя 'cc' для вызова C-компилятора, поэтому
-# создадим символическую ссылку cc -> gcc в /usr/bin/
-ln -sv gcc /usr/bin/cc
+# создадим символическую ссылку в /usr/bin/
+#    cc -> gcc
 (
-    cd "${TMP_DIR}/usr/bin" || exit 1
-    ln -sv gcc cc
+    cd "${TMP_DIR}/usr/lib" || exit 1
+    ln -svf /usr/bin/cpp cpp
+    cd ../bin || exit 1
+    ln -svf "${PRGNAME}" cc
 )
 
-# добавим символическую ссылку в /usr/lib/bfd-plugins/
-# liblto_plugin.so ->
-#    ../../libexec/gcc/x86_64-pc-linux-gnu/${VERSION}/liblto_plugin.so
-# для совместимости, чтобы разрешить сборку программ с помощью LTO (Link Time
+# добавим символическую ссылку в  /usr/lib/bfd-plugins/ для совместимости,
+# чтобы включить сборку программ с оптимизацией компоновки LTO (Link Time
 # Optimization)
-DUMPMACHINE="$("${TMP_DIR}/usr/bin/gcc" -dumpmachine)"
-(
-    cd "${TMP_DIR}/usr/lib/bfd-plugins" || exit 1
-    ln -sfv \
-        "../../libexec/gcc/${DUMPMACHINE}/${VERSION}/liblto_plugin.so" \
-        liblto_plugin.so
-)
+#    liblto_plugin.so -> \
+    #    ../../libexec/gcc/x86_64-lfs-linux-gnu/${VERSION}/liblto_plugin.so
+DUMPMACHINE="$("${TMP_DIR}/usr/bin/${PRGNAME}" -dumpmachine)"
+ln -sfv "../../libexec/${PRGNAME}/${DUMPMACHINE}/${VERSION}/liblto_plugin.so" \
+    "${TMP_DIR}/usr/lib/bfd-plugins/"
 
 # переместим некоторые файлы
 mv -v "${TMP_DIR}/usr/lib"/*gdb.py "${TMP_DIR}/usr/share/gdb/auto-load/usr/lib"
 
 chmod 755 "${TMP_DIR}/usr/lib/libgcc_s.so"{,.1}
 
-# удалим директории и файлы, которые были установлены в систему временным GCC
-rm -rf "/usr/include/c++/${VERSION}/x86_64-lfs-linux-gnu"
-rm -rf /usr/lib/gcc/x86_64-lfs-linux-gnu
-rm -rf /usr/libexec/gcc/x86_64-lfs-linux-gnu
-rm -f  /usr/bin/x86_64-lfs-linux-gnu-*
-rm -f  /usr/x86_64-lfs-linux-gnu
+# если мы устанавливаем пакет в первый раз, удалим директории и файлы, которые
+# были установлены GCC, построенным во временной системе
+if [ -d /usr/x86_64-lfs-linux-gnu ]; then
+    rm -f  /usr/bin/x86_64-lfs-linux-gnu-*
+    rm -rf "/usr/include/c++/${VERSION}/x86_64-lfs-linux-gnu"
+    rm -rf /usr/lib/gcc/x86_64-lfs-linux-gnu
+    rm -rf /usr/libexec/gcc/x86_64-lfs-linux-gnu
+    rm -rf /usr/x86_64-lfs-linux-gnu
+    rm -f  /usr/lib/libstdc++.so.6.0.30-gdb.py
+fi
 
+source "${ROOT}/stripping.sh"      || exit 1
+source "${ROOT}/update-info-db.sh" || exit 1
 /bin/cp -vR "${TMP_DIR}"/* /
 
 cat << EOF > "/var/log/packages/${PRGNAME}-${VERSION}"
@@ -119,7 +114,7 @@ cat << EOF > "/var/log/packages/${PRGNAME}-${VERSION}"
 # The GCC package contains the GNU compiler collection, which includes the C
 # and C++ compilers.
 #
-# Home page: https://gcc.gnu.org/
+# Home page: https://${PRGNAME}.gnu.org/
 # Download:  http://ftp.gnu.org/gnu/${PRGNAME}/${PRGNAME}-${VERSION}/${PRGNAME}-${VERSION}.tar.xz
 #
 EOF
